Practica 2/Backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
PyJWT==2.10.1
cryptography==45.0.5
bcrypt==4.3.0
email-validator==2.2.0
